requirements.txt
flask
pillow
flask_uploads
boto3
face_recognition
webcolors
